+++ app/database/migrations/2013_07_27_083301_create_sales_invoices_table.php
@@ -10,22 +10,21 @@ class CreateSalesInvoicesTable extends Migration {
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('sales_invoices', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('so_no')->nullable();
-			$table->integer('client')->nullable();
-			$table->string('terms')->nullable();
-			$table->float('discount')->nullable();
-			$table->float('subtotal')->nullable();
-			$table->float('vat')->nullable();
-			$table->float('total')->nullable();
-			$table->integer('created_by')->nullable();
-			$table->integer('approved_by')->nullable();
-			$table->string('status')->nullable();
-            $table->timestamps();
-        });
+                    $table->increments('id');
+                    $table->integer('so_no')->nullable();
+                    $table->integer('client')->nullable();
+                    $table->string('terms')->nullable();
+                    $table->decimal('discount', 30, 2)->nullable();
+                    $table->decimal('subtotal', 30, 2)->nullable();
+                    $table->decimal('vat', 30, 2)->nullable();
+                    $table->decimal('total', 30, 2)->nullable();
+                    $table->integer('created_by')->nullable();
+                    $table->integer('approved_by')->nullable();
+                    $table->string('status')->nullable();
+                    $table->timestamps();
+                });
     }
 
     /**
@@ -33,8 +32,7 @@ class CreateSalesInvoicesTable extends Migration {
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::drop('sales_invoices');
     }
